src/admin/pyproject.toml
[project]
name = "aithena-admin"
version = "0.1.0"
description = "Aithena Admin Dashboard"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.51.0,<2",
    "redis==4.6.0",
    "python-dotenv==1.0.1",
    "requests>=2.31.0",
    "pandas==2.2.2",
    "pillow>=12.1.1",
    "pyjwt>=2.10.0",
    "python-json-logger>=4.0.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.0",
]
